handle error on product handler and my progression of stock CRUD

+++ Entities/Stock.go
@@ -4,7 +4,7 @@ import "time"
 
 type Stock struct {
 	SID       string
-	SQuantity int64
+	SQuantity int8
 	SUpdated  time.Time
 }
